--- lambda/.eslintrc
@@ -1,11 +1,0 @@
-{
-  "parser": "@typescript-eslint/parser",
-  "plugins": ["@typescript-eslint"],
-  "extends": ["plugin:@typescript-eslint/recommended"],
-  "env": {
-    "node": true
-  },
-  "rules": {
-    "@typescript-eslint/no-unused-vars": "error"
-  }
-}
