+++ src/i18n/en.ts
@@ -1,0 +1,3 @@
+export default {
+  SKILL_NAME: 'Hello world'
+}
