+++ src/i18n/index.ts
@@ -1,0 +1,10 @@
+import en from './en'
+
+export default {
+  lng: 'en',
+  resources: {
+    en: {
+      translation: en
+    }
+  }
+}
